+++ CHANGELOG.md
@@ -29,6 +29,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - **dropdown**: Fix disabled attribute that did not work
 - **dropdown**: Dropdown overlay closes when click on another dropdown instance
 - **grid**: initial sort order attribute
+- **input**: add disabled styles
 
 ### Security